+++ packages/editor/web/src/index.ts
@@ -6,7 +6,6 @@ export {
   createEmpty,
   createWithContent,
   convertToRaw,
-  __convertToRawWithoutVersion,
   convertFromRaw,
 } from '../lib/editorStateConversion';