+++ packages/plugin-giphy/web/src/typeMapper.ts
@@ -1,5 +1,5 @@
 import GiphyViewer from './giphy-viewer';
-import { GIPHY_TYPE } from './constants';
+import { GIPHY_TYPE } from './types';
 import { PluginTypeMapper } from 'wix-rich-content-common';
 
 export const typeMapper: PluginTypeMapper = () => ({