+++ packages/plugin-image/web/src/toolbar/insert-buttons.tsx
@@ -1,14 +1,15 @@
 import { TOOLBARS, BUTTON_TYPES, INSERT_PLUGIN_BUTTONS } from 'wix-rich-content-editor-common';
 import { DEFAULTS } from '../consts';
 import { InsertPluginIcon } from '../icons';
-import { CreateInsertButtons, TranslationFunction, PluginConfig } from 'wix-rich-content-common';
+import { CreateInsertButtons, TranslationFunction } from 'wix-rich-content-common';
+import { ImagePluginEditorConfig } from '../types';
 
 const createInsertButtons: CreateInsertButtons = ({
   t,
   settings,
 }: {
   t: TranslationFunction;
-  settings: PluginConfig;
+  settings: ImagePluginEditorConfig;
 }) => {
   const icon = settings?.toolbar?.icons?.InsertPluginButtonIcon || InsertPluginIcon;
   return [