+++ packages/plugin-line-spacing/web/src/createLineSpacingPlugin.ts
@@ -1,10 +1,10 @@
 import { createBasePlugin } from 'wix-rich-content-plugin-commons';
-import { LINE_SPACING_TYPE } from './types';
+import { LINE_SPACING_TYPE, LineSpacingPluginEditorConfig } from './types';
 import createLineSpacingToolbar from './toolbar/createToolbar';
 import { CreatePluginFunction } from 'wix-rich-content-common';
 import { DEFAULTS } from './defaults';
 
-const createLineSpacingPlugin: CreatePluginFunction = config => {
+const createLineSpacingPlugin: CreatePluginFunction<LineSpacingPluginEditorConfig> = config => {
   const type = LINE_SPACING_TYPE;
   const { theme, isMobile, t, [type]: settings = {}, ...rest } = config;