+++ packages/plugin-spoiler/web/src/createSpoilerPlugin.ts
@@ -1,11 +1,11 @@
 import createToolbar from './toolbar/createToolbar';
-import { SPOILER_TYPE } from './types';
+import { SPOILER_TYPE, SpoilerPluginEditorConfig } from './types';
 import { createBasePlugin } from 'wix-rich-content-plugin-commons';
 import { styleFnFilter } from './spoilerUtilsFn';
 import { CreatePluginFunction } from 'wix-rich-content-common';
 import { DEFAULTS } from './defaults';
 
-const createSpoilerPlugin: CreatePluginFunction = config => {
+const createSpoilerPlugin: CreatePluginFunction<SpoilerPluginEditorConfig> = config => {
   const { helpers, t, [SPOILER_TYPE]: settings = {}, isMobile, ...rest } = config;
 
   return createBasePlugin({