+++ packages/plugin-spoiler/web/src/viewer.ts
@@ -1,13 +1,14 @@
-import { SPOILER_TYPE } from './types';
-import inlineStyleMapper from './spoilerInlineStyleMapper';
+import { SPOILER_TYPE, SpoilerPluginViewerConfig } from './types';
+import { spoilerInlineStyleMapper } from './spoilerInlineStyleMapper';
 import { initSpoilersContentState } from './spoilerUtilsFn';
-export { inlineStyleMapper as spoilerInlineStyleMapper, initSpoilersContentState, SPOILER_TYPE };
+import { ViewerPluginCreator } from 'wix-rich-content-common';
+export { spoilerInlineStyleMapper, initSpoilersContentState, SPOILER_TYPE };
 export { default as SpoilerViewer } from './spoiler-viewer';
 
-export const pluginSpoiler = (config = {}) => {
+export const pluginSpoiler: ViewerPluginCreator<SpoilerPluginViewerConfig> = config => {
   return {
     config: { initSpoilersContentState, ...config },
     type: SPOILER_TYPE,
-    inlineStyleMapper,
+    inlineStyleMapper: spoilerInlineStyleMapper,
   };
 };