+++ packages/plugin-video/web/src/createVideoPlugin.ts
@@ -1,10 +1,10 @@
 import createToolbar from './toolbar/createToolbar';
 import { Component, DEFAULTS } from './video-component';
-import { VIDEO_TYPE, VIDEO_TYPE_LEGACY } from './types';
+import { VIDEO_TYPE, VIDEO_TYPE_LEGACY, VideoPluginEditorConfig } from './types';
 import { createBasePlugin } from 'wix-rich-content-plugin-commons';
 import { CreatePluginFunction } from 'wix-rich-content-common';
 
-const createVideoPlugin: CreatePluginFunction = config => {
+const createVideoPlugin: CreatePluginFunction<VideoPluginEditorConfig> = config => {
   const { helpers, t, [VIDEO_TYPE]: settings = {}, isMobile, commonPubsub, ...rest } = config;
 
   return createBasePlugin({