+++ packages/template-atomic-plugin/web/src/types.ts
@@ -1,2 +1,7 @@
-const YOUR_PLUGIN_NAME_TYPE = 'yourPluginName';
-export { YOUR_PLUGIN_NAME_TYPE };
+import { EditorPluginConfig, ViewerPluginConfig } from 'wix-rich-content-common';
+
+export const YOUR_PLUGIN_NAME_TYPE = 'yourPluginName';
+
+export interface YourPluginNamePluginEditorConfig extends EditorPluginConfig {}
+
+export interface YourPluginNamePluginViewerConfig extends ViewerPluginConfig {}