+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 ### :bug: Bug Fix
 - `common`
   - [#1761](https://github.com/wix-incubator/rich-content/pull/1761) customStyles: keep `.text` color as `unset` if custom paragraph color was not defined.
+- `accordion`
+  - [#1759](https://github.com/wix-incubator/rich-content/pull/1759) fix accordion's focus when editor is not focused
 
 ### :house: Internal
 - `editor`