+++ CHANGELOG.md
@@ -16,6 +16,11 @@
     Click to see more.
   </summary>
 
+
+### :rocket: New Feature
+- `ricos-common`
+  - [#1751](https://github.com/wix-incubator/rich-content/pull/1751) pasted headings will be converted to supported headings
+
 ### :bug: Bug Fix
 - `common`
   - [#1761](https://github.com/wix-incubator/rich-content/pull/1761) customStyles: keep `.text` color as `unset` if custom paragraph color was not defined.